tsz-server: persistent node worker for native TypeScript operations
Rename/classification/etc. delegates to real \`tsc\` via a Node.js
subprocess. Previously every call spawned a fresh node process that
had to re-\`require(\"typescript\")\` — roughly 1–2 s of TS module load
amortized across every test's native-LS round-trip.

Swap in a long-lived NativeTsWorker:
- Split the embedded worker JS out to \`native_ts_worker.js\` so both
  the legacy single-shot path and the persistent path share source.
- Add a \`TSZ_NATIVE_TS_PERSISTENT=1\` mode to the script: read
  newline-delimited JSON requests from stdin, write newline-delimited
  JSON responses to stdout, and keep the TypeScript runtime hot
  across requests.
- \`Server\` grows a \`native_ts_worker: Option<Mutex<NativeTsWorker>>\`
  field populated on server startup. try_native_typescript_operation
  now tries the worker first and falls back to the original
  spawn-per-call path when the worker is unhealthy.

Effect on the fourslash suite: tests that made ~10 native-LS calls
in series (e.g. renameImport, renameTemplateLiteralsComputedProperties,
autoImportFileExcludePatterns*) now finish well under the 15 s watchdog.
Local full-suite run: 6562/6562 (100%, no timeouts).
